A lot of stuff
Main things:
I've started using illuminance for light emmiters. It's still very much a WIP and there are a few things I'll need to add to do it properly, but it's a start.

I've also separated the sky into composite. Ideally, it would be rendered before the gbuffers and readable in gbuffers_water, but I don't expect that to be possible any time soon. But I may be able to move it into gbuffers_skybasic at some point, so I am going to keep it separate, at least for now.