Added persistent time, general cleanup
The persistent time is done by using sunAngle and moonPhase in combination, which allows for time-based effects to run smoothly while still depending on the actual in-game time, which means that it will be persistent across sessions. This also lasts a bit over 2.6 hours before resetting, which is much better than the one hour you get from frameTimeCounter.