A lot of stuff
Rough reflections with multiple samples and bounces, and an option to turn them off (set samples to 0). Water still has reflections if you do turn them off, as otherwise it would look terrible.
A system for water waves, and new water waves to go with that. May do some changes to it later.
Changed how reflectance is handled, it's now RGB with fresnel applying to each of the components separately. Currently still uses the same data from the textures as before, this will eventually be changed tough.