Renamed illuminance.glsl to lightingConstants.glsl
I'll be using it for any constants related to lighting in the future. The new values in it are for what I'm doing with the sky.

I've also done a few small optimizations while updating some stuff to use the new values, not that performance was actually affected by that.