Skylight brightness based on global light, PCSS
This PCSS looks a bit better than what I was using before (which was removed because it wasn't very good), haven't really optimized it tough.

Probably won't optimize it much any time soon tough, it still has reasonable performance, and the shaderpack as a whole is getting a bit messy so I'll probably clean that up first.